+++ src/expression/embeddedDocs/function/matrix/count.js
@@ -1,0 +1,17 @@
+export const countDocs = {
+  name: 'count',
+  category: 'Matrix',
+  syntax: [
+    'count(x)'
+  ],
+  description: 'Count the number of elements of a matrix, array or string.',
+  examples: [
+    'a = [1, 2; 3, 4; 5, 6]',
+    'count(a)',
+    'size(a)',
+    'count("hello world")'
+  ],
+  seealso: [
+    'size'
+  ]
+}